prompt before delete and list project contents when passed project name

+++ ddsc/ddsclient.py
@@ -1,5 +1,6 @@
 """ Runs the appropriate command for a user based on arguments. """
 from __future__ import print_function
+import sys
 import datetime
 import pipes
 from ddsc.core.handover import ProjectHandover, HandoverError
@@ -7,6 +8,7 @@ from ddsc.core.remotestore import RemoteStore
 from ddsc.core.upload import ProjectUpload
 from ddsc.cmdparser import CommandParser, path_does_not_exist_or_is_empty, replace_invalid_path_chars
 from ddsc.core.download import ProjectDownload
+from ddsc.core.util import ProjectFilenameList
 
 NO_PROJECTS_FOUND_MESSAGE = 'No projects found.'
 
@@ -230,7 +232,7 @@ class HandoverCommand(object):
 
 class ListCommand(object):
     """
-    Print out a list of project names one line at a time.
+    Print out a list of project names one line at a time or details about a single project.
     Names are escaped so the output can be used with the delete command.
     """
     def __init__(self, config):
@@ -245,6 +247,19 @@ class ListCommand(object):
         Lists project names.
         :param args Namespace arguments parsed from the command line
         """
+        if args.project_name:
+            project = self.remote_store.fetch_remote_project(args.project_name, must_exist=True)
+            self.print_project_details(project)
+        else:
+            self.print_project_names()
+
+    def print_project_details(self, project):
+        filename_list = ProjectFilenameList()
+        filename_list.walk_project(project)
+        for info in filename_list.details:
+            print(info)
+
+    def print_project_names(self):
         names = self.remote_store.get_project_names()
         if names:
             for name in names:
@@ -263,11 +278,31 @@ class DeleteCommand(object):
         :param config: Config global configuration for use with this command.
         """
         self.remote_store = RemoteStore(config)
-        self.project_handover = ProjectHandover(config, self.remote_store, print_func=print)
 
     def run(self, args):
         """
         Deletes a single project specified by project_name in args.
         :param args Namespace arguments parsed from the command line
         """
-        self.remote_store.delete_project_by_name(args.project_name)
+        project_name = args.project_name
+        project = self.remote_store.fetch_remote_project(project_name, must_exist=False)
+        if not project:
+            raise ValueError("No project named '{}' found.\n".format(project_name))
+        else:
+            if not args.force:
+                delete_prompt = "Are you sure you wish to delete {} (y/n)?".format(project_name)
+                if not boolean_input_prompt(delete_prompt):
+                    return
+            self.remote_store.delete_project_by_name(args.project_name)
+
+
+def boolean_input_prompt(message):
+    if sys.version_info >= (3, 0, 0):
+        result = input(message)
+    else:
+        result = raw_input(message)
+    result = result.upper()
+    return result == "Y" or result == "YES" or result == "T" or result == "TRUE"
+
+
+
